Fix parameter type in request_available_modes function for correct mode index handling

+++ src/mavsdk/plugins/striker/striker_impl.cpp
@@ -386,7 +386,7 @@ void StrikerImpl::request_available_modes(uint32_t modeIndex)
 
     command.command = MAV_CMD_REQUEST_MESSAGE;
     command.params.maybe_param1 = MAVLINK_MSG_ID_AVAILABLE_MODES;
-    command.params.maybe_param2 = static_cast<float>(modeIndex);
+    command.params.maybe_param2 = modeIndex;
 
     _system_impl->send_command_async(command, [this](MavlinkCommandSender::Result result, float) {
         if (result == MavlinkCommandSender::Result::Success) {
